+++ service/controller/resource/monitoring/remotewritesecret/create.go
@@ -1,44 +1,35 @@
-package remotewriteapiendpointconfigsecret
+package remotewritesecret
 
 import (
 	"context"
 	"reflect"
 
 	"github.com/giantswarm/microerror"
+	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/yaml"
 
 	remotewriteconfiguration "github.com/giantswarm/prometheus-meta-operator/v2/pkg/remotewrite/configuration"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
 )
 
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	// If we are in a management cluster, we do not need this secret
-	if key.IsManagementCluster(r.Installation, obj) {
-		return r.EnsureDeleted(ctx, obj)
-	}
-
-	r.logger.Debugf(ctx, "ensuring prometheus remote write api endpoint secret")
+	r.logger.Debugf(ctx, "ensuring prometheus remote write secret")
 	{
+
 		cluster, err := key.ToCluster(obj)
 		if err != nil {
 			return microerror.Mask(err)
 		}
 
-		// Get password from remote-write-secret
-		r.logger.Debugf(ctx, "looking up for secret remote write secret")
-		_, password, err := remotewriteconfiguration.GetUsernameAndPassword(r.k8sClient.K8sClient(), ctx, cluster, r.Installation, r.Provider)
-		if err != nil {
-			r.logger.Errorf(ctx, err, "lookup for remote write secret failed")
-			return microerror.Mask(err)
-		}
-
-		name := key.RemoteWriteAPIEndpointConfigSecretName(cluster, r.Installation)
+		name := key.RemoteWriteSecretName(cluster)
 		namespace := key.GetClusterAppsNamespace(cluster, r.Installation, r.Provider)
+
 		// Get the current secret if it exists.
 		current, err := r.k8sClient.K8sClient().CoreV1().Secrets(namespace).Get(ctx, name, metav1.GetOptions{})
 		if apierrors.IsNotFound(err) {
-			err = r.createSecret(ctx, cluster, name, namespace, password, r.Version)
+			err = r.createSecret(ctx, cluster, name, namespace)
 			if err != nil {
 				return microerror.Mask(err)
 			}
@@ -47,7 +38,14 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		}
 
 		if current != nil {
-			desired, err := r.desiredSecret(cluster, name, namespace, password, r.Version)
+			// As it takes a long time to apply the new password to the agent due to a built-in delay in the app-platform,
+			// we keep the already generated remote write password.
+			password, err := readRemoteWritePasswordFromSecret(*current)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+
+			desired, err := r.desiredSecret(cluster, name, namespace, password)
 			if err != nil {
 				return microerror.Mask(err)
 			}
@@ -61,9 +59,25 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		}
 	}
 
-	r.logger.Debugf(ctx, "ensured prometheus remote write api endpoint secret")
+	r.logger.Debugf(ctx, "ensured prometheus remote write secret")
 
 	return nil
+}
+
+func readRemoteWritePasswordFromSecret(secret corev1.Secret) (string, error) {
+	remoteWriteConfig := remotewriteconfiguration.RemoteWriteConfig{}
+	err := yaml.Unmarshal(secret.Data["values"], &remoteWriteConfig)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+
+	for _, rw := range remoteWriteConfig.PrometheusAgentConfig.RemoteWrite {
+		if rw.Name == key.PrometheusMetaOperatorRemoteWriteName {
+			return rw.Password, nil
+		}
+	}
+
+	return "", remoteWriteNotFound
 }
 
 func updateMeta(c, d metav1.Object) {