+++ service/controller/resource/monitoring/remotewritesecret/error.go
@@ -1,17 +1,8 @@
-package remotewriteingressauth
+package remotewritesecret
 
 import (
 	"github.com/giantswarm/microerror"
 )
-
-var secretNotFound = &microerror.Error{
-	Kind: "secretNotFound",
-}
-
-// IsSecretNotFound asserts secretNotFound.
-func IsSecretNotFound(err error) bool {
-	return microerror.Cause(err) == secretNotFound
-}
 
 var remoteWriteNotFound = &microerror.Error{
 	Kind: "remoteWriteNotFound",